requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.0.0
django-taggit==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
python-decouple==3.8
